Simplify use of cache

+++ src/cache.sh
@@ -1,21 +1,23 @@
 function prepare_cache () {
-	expect_vars HALCYON_CACHE_DIR HALCYON_PURGE_CACHE
+	expect_vars HALCYON_CACHE_DIR HALCYON_RECURSIVE HALCYON_PURGE_CACHE HALCYON_NO_CACHE
 
 	local cache_dir
 	expect_args cache_dir -- "$@"
 
+	if (( HALCYON_RECURSIVE )) || (( HALCYON_NO_CACHE )); then
+		return 0
+	fi
+
 	if (( HALCYON_PURGE_CACHE )); then
 		log 'Purging cache'
+		log
 
 		rm -rf "${HALCYON_CACHE_DIR}"
 	fi
 
-	mkdir -p "${HALCYON_CACHE_DIR}" || die
-	rm -f "${HALCYON_CACHE_DIR}/.halcyon-mark" || die
+	mkdir -p "${HALCYON_CACHE_DIR}" "${cache_dir}" || die
 
 	if ! (( HALCYON_PURGE_CACHE )); then
-		log 'Examining cache contents'
-
 		local files
 		if files=$(
 			find_tree "${HALCYON_CACHE_DIR}" -maxdepth 1 -type f 2>'/dev/null' |
@@ -23,42 +25,44 @@ function prepare_cache () {
 			sort_naturally |
 			match_at_least_one
 		); then
+			log 'Examining cache contents'
+
 			tar_copy "${HALCYON_CACHE_DIR}" "${cache_dir}" || die
 
 			quote <<<"${files}"
-		else
-			log_indent '(none)'
+			log
 		fi
 	fi
 
-	touch "${HALCYON_CACHE_DIR}/.halcyon-mark" || die
+	touch "${cache_dir}" || die
 }
 
 
 function clean_cache () {
-	expect_vars HALCYON_CACHE_DIR
-	expect_existing "${HALCYON_CACHE_DIR}/.halcyon-mark"
+	expect_vars HALCYON_CACHE_DIR HALCYON_RECURSIVE HALCYON_NO_CACHE
 
 	local cache_dir
 	expect_args cache_dir -- "$@"
 
+	if (( HALCYON_RECURSIVE )) || (( HALCYON_NO_CACHE )); then
+		return 0
+	fi
+
 	local mark_time name_prefix
-	mark_time=$( get_file_modification_time "${HALCYON_CACHE_DIR}/.halcyon-mark" ) || die
+	mark_time=$( get_modification_time "${cache_dir}" ) || die
 	name_prefix=$( format_sandbox_constraints_file_name_prefix ) || die
 
-	rm -f "${HALCYON_CACHE_DIR}/.halcyon-mark" "${HALCYON_CACHE_DIR}/${name_prefix}"* || die
+	rm -f "${HALCYON_CACHE_DIR}/${name_prefix}"* || die
 
 	local file
 	find "${HALCYON_CACHE_DIR}" -maxdepth 1 -type f 2>'/dev/null' |
 		while read -r file; do
 			local file_time
-			file_time=$( get_file_modification_time "${file}" ) || die
+			file_time=$( get_modification_time "${file}" ) || die
 			if (( file_time < mark_time )); then
 				rm -f "${file}" || die
 			fi
 		done
-
-	log 'Examining cache changes'
 
 	local changed_files
 	if changed_files=$(
@@ -66,8 +70,9 @@ function clean_cache () {
 		filter_not_matching '^= ' |
 		match_at_least_one
 	); then
+		log
+		log 'Examining cache changes'
+
 		quote <<<"${changed_files}"
-	else
-		log_indent '(none)'
 	fi
 }
